feat(core): 优化验证码功能并添加用户信息相关接口
- 修改验证码生成和验证逻辑，支持不同类型的验证码
- 优化用户登录和获取当前用户信息的接口- 调整角色和路由相关的实体类和DTO
-修复部分已知问题并优化系统性能

+++ services/auth/src/main/java/com/yuan/cloud/auth/vo/YaCaptchaVO.java
@@ -30,6 +30,6 @@ public class YaCaptchaVO {
     /**
      * 验证码过期时间
      */
-    private Long captchaExpireAt;
+    private int captchaExpireAt;
 
 }
